[DQMOffline/Trigger] moved more headers into plugins/*.cc
